+++ src/Vahana.jl
@@ -39,4 +39,5 @@ include("PrettyPrinting.jl")
 include("REPL.jl")
 
 include("GraphsSupport.jl")
+include("Raster.jl")
 end